Rename loader.h/.cpp to shader_loader.h/.cpp
