+++ vision/main.cpp
@@ -13,7 +13,7 @@ using namespace glm;
 #include <map>
 #include <fstream>
 
-#include "loader.h"
+#include "shader_loader.h"
 
 void GLAPIENTRY MessageCallback( GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam ){fprintf( stderr, "GL CALLBACK: %s type = 0x%x, severity = 0x%x, message = %s\n", ( type == GL_DEBUG_TYPE_ERROR ? "** GL ERROR **" : "" ), type, severity, message );}
 static void glfwError(int id, const char* description){std::cout << description << std::endl;}
